# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Condominium",
        "description": "API documentation for the Condominium management system",
        "version": "1.0"
    },
    "paths": {
...
        "/api/positions": {
            "get": {
                "tags": [
                    "Positions"
                ],
                "summary": "Listar todos os cargos",
                "description": "Retorna uma lista de cargos paginados",
                "operationId": "979e46c0161e2570ef5478434095cb31",
                "responses": {
                    "200": {
                        "description": "Lista de cargos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Manager"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-08T12:34:56Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-08T12:34:56Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de cargos retornada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Positions"
                ],
                "summary": "Criar um novo cargo",
                "description": "Cria um cargo e retorna os detalhes do novo cargo",
                "operationId": "8b47544f5112eedfa8e15bfa7d23ba46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cargo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cargo criado com sucesso!"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Manager"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-08T12:34:56Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-08T12:34:56Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Já existe um cargo com esse nome!"
                    },
                    "500": {
                        "description": "Erro ao criar um cargo"
                    }
                }
            }
        },
        "/api/positions/{id}": {
            "get": {
                "tags": [
                    "Positions"
                ],
                "summary": "Exibir detalhes de um cargo",
                "description": "Retorna os detalhes de um cargo específico",
                "operationId": "02987e1d2529f90862c1e6ebeed617a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cargo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Manager"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-08T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-08T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cargo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Positions"
                ],
                "summary": "Atualizar um cargo",
                "description": "Atualiza as informações de um cargo específico",
                "operationId": "e01f80e1ad1b8b3f65a959aaf8b1f4d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cargo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Manager"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cargo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cargo atualizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cargo não encontrado"
                    },
                    "500": {
                        "description": "Erro ao atualizar um cargo"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Positions"
                ],
                "summary": "Deletar um cargo",
                "description": "Deleta um cargo específico",
                "operationId": "be998660e7df336ea7e1b510aefe2694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do cargo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cargo deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cargo deletado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cargo não encontrado"
                    },
                    "500": {
                        "description": "Erro ao deletar o cargo"
                    }
                }
            }
        }
    },
...
}